pyproject.toml
[project]
name = "g-seek"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv",
    "langchain-community",
    "langchain-openai",
    "langchain-ollama",
    "duckduckgo-search",
    "streamlit",
    "pyvis",
    "beautifulsoup4",
    "requests",
    "pypdf",
    "streamlit-agraph",
    "ddgs>=9.10.0",
    "chainlit",
    "langchain>=1.2.6",
    "langchain-core>=1.2.7",
    "pydantic-settings>=2.1.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
